% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{api_request}
\alias{api_request}
\title{Make an API GET-request}
\usage{
api_request(url = NULL, apikey = NULL, cmd = "get_servers_info", ...)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{cmd}{API command,
see \href{https://github.com/Tautulli/Tautulli/blob/master/API.md}{the API docs}. Defaults to
printing server information via the \code{get_servers_info} method.}

\item{...}{Optional (named) parameters.}
}
\value{
The API result, usually a \code{list}.
}
\description{
Make an API GET-request
}
\examples{
\dontrun{
api_request("http://example.com/plexpy", "asdf", "get_servers_info")
}
}
