% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{arnold}
\alias{arnold}
\title{Arnold}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#arnold}
}
\usage{
arnold(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{character} of length \code{1}.
}
\description{
Arnold
}
\examples{
\dontrun{
arnold()
}
}
