% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity.R
\name{get_activity}
\alias{get_activity}
\title{Get Current Server Activity}
\usage{
get_activity(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{list} with bandwidth + stream info and a session \code{tbl}.
}
\description{
Get Current Server Activity
}
\examples{
\dontrun{
get_activity()
}
}
