% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history.R
\name{get_history}
\alias{get_history}
\title{Title}
\usage{
get_history(url = NULL, apikey = NULL, grouping = 0, user = NULL,
  user_id = NULL, rating_key = NULL, parent_rating_key = NULL,
  grandparent_rating_key = NULL, start_date = NULL, section_id = NULL,
  media_type = NULL, transcode_decision = NULL, order_column = NULL,
  order_dir = "desc", start = 0, length = 25, search = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{grouping}{0 (default) or 1}

\item{user}{"Jon Snow"}

\item{user_id}{133788}

\item{rating_key}{4348}

\item{parent_rating_key}{544}

\item{grandparent_rating_key}{351}

\item{start_date}{"YYYY-MM-DD"}

\item{section_id}{2}

\item{media_type}{"movie", "episode", "track"}

\item{transcode_decision}{"direct play", "copy", "transcode",}

\item{order_column}{"date", "friendly_name", "ip_address", "platform", "player", "full_title", "started", "paused_counter", "stopped", "duration"}

\item{order_dir}{"desc" (default) or "asc"}

\item{start}{Row to start from, 0 (default)}

\item{length}{Number of items to return, 25}

\item{search}{A string to search for, "Thrones"}
}
\value{
A \code{list} with totals and the history as \code{data.frame}
}
\description{
Title
}
\examples{
\dontrun{
get_history(length = 10)
}
}
