% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home_stats.R
\name{get_home_stats}
\alias{get_home_stats}
\title{Get Home Stats}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_home_stats}
}
\usage{
get_home_stats(
  url = NULL,
  apikey = NULL,
  grouping = 0,
  time_range = 30,
  stats_type = 0,
  stats_count = 5
)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{grouping}{0 or 1}

\item{time_range}{The time range to calculate statistics, default is \code{30}}

\item{stats_type}{\code{0} for plays (default), \code{1} for duration}

\item{stats_count}{The number of top items to list, default is \code{5}}
}
\value{
A \code{list} of length 10, with \code{tbl}s for each category of stats.
}
\description{
The data displayed on the \emph{Tautulli} home.
}
\examples{
\dontrun{
get_home_stats()
}
}
