% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library.R
\name{get_libraries}
\alias{get_libraries}
\title{Get All the Libraries' Data}
\usage{
get_libraries(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{tbl} with columns \code{section_id} and \code{section_name}
}
\description{
Get All the Libraries' Data
}
\examples{
\dontrun{
get_library(section_id = 1)
}
}
