% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library.R
\name{get_libraries_table}
\alias{get_libraries_table}
\title{Get All the Libraries-Table}
\usage{
get_libraries_table(url = NULL, apikey = NULL, order_column = NULL,
  order_dir = "desc", start = 0, length = 25, search = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{order_column}{Optional, "added_at", "title", "container", "bitrate", "video_codec",
"video_resolution", "video_framerate", "audio_codec", "audio_channels",
"file_size", "last_played", "play_count"}

\item{order_dir}{"desc" (default) or "asc"}

\item{start}{Row to start from, default 0}

\item{length}{Number of items to return, default 25}

\item{search}{A string to search for}
}
\value{
A \code{tbl} with columns \code{section_id} and \code{section_name}
}
\description{
Get All the Libraries-Table
}
\examples{
\dontrun{
get_libraries_table()
}
}
