% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library.R
\name{get_library}
\alias{get_library}
\title{Get A Single Library Information}
\usage{
get_library(url = NULL, apikey = NULL, section_id)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{section_id}{The library's \code{section_id}, see \link{get_library_names}}
}
\value{
A \code{tbl} with columns \code{section_id} and \code{section_name}
}
\description{
Get A Single Library Information
}
\examples{
\dontrun{
get_library(section_id = 1)
}
}
