% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library.R
\name{get_library_names}
\alias{get_library_names}
\title{Get Library Sections and Names}
\usage{
get_library_names(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{data.frame} with columns \code{section_id} and \code{section_name}
}
\description{
Get Library Sections and Names
}
\examples{
\dontrun{
get_library_names()
}
}
