% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library.R
\name{get_library_user_stats}
\alias{get_library_user_stats}
\title{Get Library User Stats}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_library_user_stats}
}
\usage{
get_library_user_stats(url = NULL, apikey = NULL, section_id)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{section_id}{The library's \code{section_id}, e.g. \code{1}.}
}
\value{
A \code{tbl} with 4 columns and one row per user.
}
\description{
Get Library User Stats
}
\examples{
\dontrun{
get_library_user_stats(section_id = 1)
}
}
