% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library.R
\name{get_library_watch_time_stats}
\alias{get_library_watch_time_stats}
\title{Get Library Watch Time Stats}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_library_watch_time_stats}
}
\usage{
get_library_watch_time_stats(url = NULL, apikey = NULL, section_id = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{section_id}{The id of the Plex library section, OR}
}
\value{
A \code{tbl} of length 3
}
\description{
Get Library Watch Time Stats
}
\examples{
\dontrun{
get_library_watch_time_stats(section_id = 2)
}
}
