% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by.R
\name{get_plays_by_date}
\alias{get_plays_by_date}
\title{Get Plays by Date}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_plays_by_date}
}
\usage{
get_plays_by_date(url = NULL, apikey = NULL, time_range = 30,
  y_axis = "plays", user_id = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{time_range}{The number of days of data to return, default is \code{30}.}

\item{y_axis}{\code{"plays"} (default) or \code{"duration"}}

\item{user_id}{The user id to filter the data}
}
\value{
A \code{tbl} with columns \code{date}, \code{Movies}, \code{TV}, \code{Music}
}
\description{
Get Plays by Date
}
\examples{
\dontrun{
get_plays_by_date(y_axis = "duration")
}
}
\seealso{
Other Playback History: \code{\link{get_plays_by_dayofweek}},
  \code{\link{get_plays_by_hourofday}},
  \code{\link{get_plays_by_source_resolution}},
  \code{\link{get_plays_by_stream_resolution}},
  \code{\link{get_plays_by_stream_type}},
  \code{\link{get_plays_per_month}}
}
