% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top10.R
\name{get_plays_by_top_10_platforms}
\alias{get_plays_by_top_10_platforms}
\title{Get Plays by Top 10 Platforms}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_plays_by_top_10_platforms}
}
\usage{
get_plays_by_top_10_platforms(url = NULL, apikey = NULL, time_range = 30,
  y_axis = "plays", user_id = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{time_range}{The number of days of data to return, default is \code{30}.}

\item{y_axis}{\code{"plays"} (default) or \code{"duration"}}

\item{user_id}{The user id to filter the data}
}
\value{
A \code{tbl} with columns \code{hour}, \code{Movies}, \code{TV}, \code{Music}
}
\description{
Get Plays by Top 10 Platforms
}
\examples{
\dontrun{
get_plays_by_top_10_platforms()
}
}
