% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pms.R
\name{get_pms_update}
\alias{get_pms_update}
\title{Check for PMS Updates}
\usage{
get_pms_update(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{list} containing update information
}
\description{
Check for PMS Updates
}
\examples{
\dontrun{
get_pms_update()
}
}
