% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recently_added.R
\name{get_recently_added}
\alias{get_recently_added}
\title{Get Recently Added Media Items}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_recently_added}
}
\usage{
get_recently_added(url = NULL, apikey = NULL, count = 10,
  section_id = NULL, start = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{count}{The number of items to return, default is \code{10}.}

\item{section_id}{Optional. The id of the Plex library section, e.g. `1``}

\item{start}{Optional. The item number to start at, e.g. \code{0}}
}
\value{
A \code{tbl} with \code{count} rows and 18 columns.
}
\description{
Get Recently Added Media Items
}
\examples{
\dontrun{
get_recently_added(section_id = 1)
}
}
