% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servers_info.R
\name{get_server_friendly_name}
\alias{get_server_friendly_name}
\title{Get Server Friendly Name}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_server_friendly_name}
}
\usage{
get_server_friendly_name(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{tbl} with 8 columns minimum with one row per server.
}
\description{
Get Server Friendly Name
}
\examples{
\dontrun{
get_server_friendly_name()
}
}
\seealso{
Other Server Information: 
\code{\link{get_server_identity}()},
\code{\link{get_server_id}()},
\code{\link{get_server_list}()},
\code{\link{get_servers_info}()}
}
\concept{Server Information}
