% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servers_info.R
\name{get_server_id}
\alias{get_server_id}
\title{Get Server ID}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_server_id}
}
\usage{
get_server_id(url = NULL, apikey = NULL, hostname, port = 32400,
  remote = 0, ssl = 0)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{hostname}{THe hostname}

\item{port}{The server port, default is \code{32400}}

\item{ssl, remote}{Optional, 0 for no, 1 for yes}
}
\value{
A \code{tbl} with 8 columns minimum with one row per server.
}
\description{
Get Server ID
}
\examples{
\dontrun{
get_server_id()
}
}
\seealso{
Other Server Information: \code{\link{get_server_friendly_name}},
  \code{\link{get_server_identity}},
  \code{\link{get_server_list}},
  \code{\link{get_servers_info}}
}
\concept{Server Information}
