% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servers_info.R
\name{get_server_identity}
\alias{get_server_identity}
\title{Get Server Identity}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_servers_identity}
}
\usage{
get_server_identity(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{tbl} with 2 columns.
}
\description{
Get Server Identity
}
\examples{
\dontrun{
get_server_identity()
}
}
\seealso{
Other Server Information: \code{\link{get_server_friendly_name}},
  \code{\link{get_server_id}},
  \code{\link{get_server_list}},
  \code{\link{get_servers_info}}
}
\concept{Server Information}
