% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servers_info.R
\name{get_server_list}
\alias{get_server_list}
\title{Get Server List}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_server_list}
}
\usage{
get_server_list(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{tbl} with 8 columns minimum with one row per server.
}
\description{
Get Server List
}
\examples{
\dontrun{
get_server_list()
}
}
\seealso{
Other Server Information: \code{\link{get_server_friendly_name}},
  \code{\link{get_server_identity}},
  \code{\link{get_server_id}},
  \code{\link{get_servers_info}}
}
\concept{Server Information}
