% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_servers_info.R
\name{get_servers_info}
\alias{get_servers_info}
\title{Get Server Information}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_servers_info}
}
\usage{
get_servers_info(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{tbl} with 5 columns minimum.
}
\description{
Get Server Information
}
\examples{
\dontrun{
get_servers_info()
}
}
\seealso{
Other Server Information: \code{\link{get_server_identity}},
  \code{\link{get_server_list}}
}
