% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top10.R
\name{get_stream_type_by_top_10_users}
\alias{get_stream_type_by_top_10_users}
\title{Get Stream Type by Top 10 Users}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_stream_type_by_top_10_users}
}
\usage{
get_stream_type_by_top_10_users(url = NULL, apikey = NULL,
  time_range = 30, y_axis = "plays", user_id = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{time_range}{The number of days of data to return, default is \code{30}.}

\item{y_axis}{\code{"plays"} (default) or \code{"duration"}}

\item{user_id}{The user id to filter the data}
}
\value{
A \code{tbl} with columns \code{user}, \code{Direct_Play}, \code{Direct_Stream}, \code{Transcode}.
}
\description{
Get Stream Type by Top 10 Users
}
\examples{
\dontrun{
get_stream_type_by_top_10_users()
}
}
