% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user_ips}
\alias{get_user_ips}
\title{Get User IPs}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_user_ips}
}
\usage{
get_user_ips(url = NULL, apikey = NULL, user_id, order_column = NULL,
  order_dir = "desc", start = 0, length = 50, search = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{user_id}{The \code{user_id} of the user.}

\item{order_column}{"user_thumb", "friendly_name", "last_seen", "ip_address", "platform",
"player", "last_played", "plays", "duration"}

\item{order_dir}{"desc" (default) or "asc"}

\item{start}{Row to start from, default is 0}

\item{length}{Number of items to return, default is 50.}

\item{search}{A string to search for}
}
\value{
A \code{tbl}
}
\description{
Get User IPs
}
\examples{
\dontrun{
get_user_ips(user_id = 192023)
}
}
