% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user_names}
\alias{get_user_names}
\title{Get User Names and IDs}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_user_names}
}
\usage{
get_user_names(url = NULL, apikey = NULL, add_pseudonym = FALSE)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{add_pseudonym}{Set to \code{TRUE} if you want to pseudonymize your user's names.}
}
\value{
A \code{tbl}
}
\description{
Get User Names and IDs
}
\examples{
\dontrun{
get_user_names()
}
}
