% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user_player_stats}
\alias{get_user_player_stats}
\title{Get a Single User's Player Stats}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_user_player_stats}
}
\usage{
get_user_player_stats(url = NULL, apikey = NULL, user_id)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{user_id}{The numeric \code{user_id}. Can be retrieved via \link{get_user_names}.}
}
\value{
A \code{tbl}
}
\description{
Get a Single User's Player Stats
}
\examples{
\dontrun{
get_user_player_stats(user_id = 1352909)
}
}
