% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_users}
\alias{get_users}
\title{Get Users}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_users}
}
\usage{
get_users(url = NULL, apikey = NULL, add_pseudonym = FALSE)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
A \code{tbl}
}
\description{
Get Users
}
\examples{
\dontrun{
get_users()
}
}
