% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_users_table}
\alias{get_users_table}
\title{Get a Users Table}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_users_table}
}
\usage{
get_users_table(url = NULL, apikey = NULL, length = 50)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{length}{Number of items to return, default is 50.}
}
\value{
A \code{tbl}
}
\description{
Get a Users Table
}
\examples{
\dontrun{
get_users_table()
}
}
