% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_server}
\alias{search_server}
\title{Search for Media Items}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#search}
}
\usage{
search_server(url = NULL, apikey = NULL, query, limit = 10)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{query}{The query string to search for}

\item{limit}{The maximum number of items to return per media type, default is \code{10}}
}
\value{
A \code{list} with each type of media
}
\description{
Search for Media Items
}
\note{
I would have named the function \code{search}, but you know, \code{base::search}.
}
\examples{
\dontrun{
search_server(query = "game of Thrones", limit = 1)
}
}
