% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauturri-package.R
\docType{package}
\name{tauturri-package}
\alias{tauturri}
\alias{tauturri-package}
\title{tauturri: Get Data Out of 'Tautulli' (Formerly 'PlexPy')}
\description{
'Tautulli' (<http://tautulli.com>) is a monitoring application for 'Plex' Media Servers
(<https://www.plex.tv>) which collects a lot of data about media items and server
usage such as play counts. This package interacts with the 'Tautulli' API of any
specified server to get said data into R. The 'Tautulli' API documentation is available
at <https://github.com/Tautulli/Tautulli/blob/master/API.md>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jemus42/tauturri}
  \item Report bugs at \url{https://github.com/jemus42/tauturri/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de}

}
\keyword{internal}
