% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tautulli.R
\name{update_check}
\alias{update_check}
\title{Check for Tautulli Updates}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/beta/API.md#update_check}
}
\usage{
update_check(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
Invisibly \code{TRUE} when an update is available, \code{FALSE} when no update is available.
}
\description{
Check for Tautulli Updates
}
\examples{
\dontrun{
update_check()
}
}
