% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tautulli.R
\name{update_tautulli}
\alias{update_tautulli}
\title{Apply Tautulli Updates}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/beta/API.md#update}
}
\usage{
update_tautulli(url = NULL, apikey = NULL)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}
}
\value{
Invisibly \code{TRUE} when the update is applied, \code{FALSE} when there is an issue.
}
\description{
Apply Tautulli Updates
}
\examples{
\dontrun{
update_tautulli()
}
}
