% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_users_table}
\alias{get_users_table}
\title{Get a Users Table}
\source{
\url{https://github.com/Tautulli/Tautulli/blob/master/API.md#get_users_table}
}
\usage{
get_users_table(
  url = NULL,
  apikey = NULL,
  order_column = NULL,
  order_dir = "desc",
  start = 0,
  length = 50,
  search = NULL
)
}
\arguments{
\item{url}{Server URL.}

\item{apikey}{API key.}

\item{order_column}{"user_thumb", "friendly_name", "last_seen", "ip_address", "platform",
"player", "last_played", "plays", "duration"}

\item{order_dir}{"desc" (default) or "asc"}

\item{start}{Row to start from, default is 0}

\item{length}{Number of items to return, default is 50.}

\item{search}{A string to search for}
}
\value{
A \code{tbl}
}
\description{
Get a Users Table
}
\examples{
\dontrun{
get_users_table()
}
}
